# Don't parse and extract streets if we don't import ways.