When importing OSM, streets were parsed and extracted regardless of whether or not they were needed.

This patch puts all the street parsing and extracting in the conditional that checks for `--import-way`. 

On imports with `osm2mimir` that don't have `--import-way` set, this greatly improves performance. Especially on OSM files with lots of streets (obviously).

Here's a rough timer ran on before and after the patch. CPU-time improved by 33%.

```
148M mei  5 02:03 /mnt/sda/OSM/mimir/osm/netherlands-gelderland-latest.osm.pbf

before:
time RUST_LOG=error ./target/release/osm2mimir --dataset=nl --input /mnt/sda/OSM/mimir/osm/netherlands-gelderland-latest.osm.pbf --poi-config ./poi_config.json --import-poi

real	1m54,401s
user	2m36,302s
sys	0m8,153s

after:
time RUST_LOG=error ./target/release/osm2mimir --dataset=nl --input /mnt/sda/OSM/mimir/osm/netherlands-gelderland-latest.osm.pbf --poi-config ./poi_config.json --import-poi

real	1m34,879s
user	1m34,825s
sys	0m5,450s
```